refactor: embed Castle webview script
